Revert "background 코드 리팩토링 및 순환참조 오류 해결"
This reverts commit 9ff3ec44c5a063dbedb92cfe5aaf9991f83ffd9b.
