+++ backend/app/api/v1/chatbot.py
@@ -9,49 +9,39 @@ from pydantic import BaseModel
 from app.api.deps import get_or_create_user
 from app.models.users import Users
 from app.utils.llm_client import LLMClient
-from app.core.prompts import SYSTEM_PROMPTS, EVENT_EXPLAIN_PROMPTS
 from app.constants import UserLevel
 
 
 chatbot_router = APIRouter()
 
+# Temporary system prompts per level
+SYSTEM_PROMPTS = {
+    UserLevel.BEGINNER: "당신은 주식 투자를 처음 접한 사용자를 돕는 친절한 조력자입니다.",
+    UserLevel.INTERMEDIATE: "당신은 투자 경험이 어느 정도 있는 사용자를 위한 조력자입니다.",
+    UserLevel.ADVANCED: "당신은 고급 투자 지식을 가진 사용자를 위한 전문 조력자입니다.",
+}
+
 
 class ChatMessage(BaseModel):
-    """채팅 메시지 모델"""
-    role: str  # "user" 또는 "assistant"
-    content: str  # 메시지 내용
+    role: str
+    content: str
 
 
 class ConversationRequest(BaseModel):
-    """사용자 대화내역 기반 질문 요청
-    
-    사용자가 챗봇과 이전에 나눈 대화 내역을 바탕으로 
-    연속적인 질문을 할 때 사용하는 요청 모델
-    """
-    history: List[ChatMessage] = []  # 이전 대화 내역
-    question: str  # 사용자의 새로운 질문
+    history: List[ChatMessage] = []
+    question: str
 
 
 class EventExplainRequest(BaseModel):
-    """특정 이벤트(금융일정)에 대한 설명 요청
-    
-    특정 금융 이벤트나 경제 지표에 대한 설명을 요청할 때 사용하는 모델
-    대화 내역 없이 단발성 질문으로 사용됨
-    """
-    event_info: str  # 설명을 요청하는 이벤트 정보
+    event_info: str
 
 
-def _build_messages(level: UserLevel, history: List[ChatMessage], question: str) -> List[dict]:
-    """대화 메시지 구성
-    
-    Args:
-        level: 사용자 레벨
-        history: 이전 대화 내역
-        question: 새로운 질문
-        
-    Returns:
-        LLM에 전달할 메시지 리스트
-    """
+class EventFollowupRequest(BaseModel):
+    history: List[ChatMessage] = []
+    question: str
+
+
+def _build_messages(level: UserLevel, history: List[ChatMessage], question: str):
     system_prompt = SYSTEM_PROMPTS.get(level, SYSTEM_PROMPTS[UserLevel.BEGINNER])
     messages = [{"role": "system", "content": system_prompt}]
     messages += [msg.dict() for msg in history]
@@ -63,11 +53,6 @@ def _build_messages(level: UserLevel, history: List[ChatMessage], question: str)
 async def conversation(
     req: ConversationRequest, db_user: Users = Depends(get_or_create_user)
 ):
-    """대화 내역 기반 질문 처리
-    
-    사용자의 이전 대화 내역을 바탕으로 연속적인 질문에 대해 스트리밍 응답을 제공합니다.
-    사용자 레벨에 따라 적절한 시스템 프롬프트를 사용합니다.
-    """
     messages = _build_messages(db_user.level, req.history, req.question)
     llm = LLMClient()
 
@@ -82,20 +67,20 @@ async def conversation(
 async def explain_event(
     req: EventExplainRequest, db_user: Users = Depends(get_or_create_user)
 ):
-    """금융 이벤트 설명
-    
-    특정 금융 이벤트나 경제 지표에 대한 설명을 스트리밍 응답으로 제공합니다.
-    사용자 레벨에 따라 적절한 깊이의 설명을 제공합니다.
-    """
-    system_prompt = EVENT_EXPLAIN_PROMPTS.get(db_user.level, EVENT_EXPLAIN_PROMPTS[UserLevel.BEGINNER])
     messages = [
-        {"role": "system", "content": system_prompt},
+        {"role": "system", "content": SYSTEM_PROMPTS.get(db_user.level)},
         {"role": "user", "content": req.event_info},
     ]
     llm = LLMClient()
+    answer = await llm.chat(messages)
+    return {"answer": answer}
 
-    async def stream():
-        async for chunk in llm.stream_chat(messages):
-            yield chunk
 
-    return StreamingResponse(stream(), media_type="text/plain")
+@chatbot_router.post("/event/followup")
+async def event_followup(
+    req: EventFollowupRequest, db_user: Users = Depends(get_or_create_user)
+):
+    messages = _build_messages(db_user.level, req.history, req.question)
+    llm = LLMClient()
+    answer = await llm.chat(messages)
+    return {"answer": answer}
