+++ backend/app/models/__init__.py
@@ -1,23 +1,8 @@
 """
 Models package for Market Timing Calendar
 """
-# Base를 먼저 import
-from .base import Base, BaseModel
-
-# 독립적인 모델들 먼저 import (외래키 관계 없는 것들)
-from .users import Users, LevelFeature
 from .events import Events, EventWebhook
+# from .users import Users, UserEventSubscription, UserGoogleCalendar, LevelFeature
+from .base import Base
 
-# 관계형 모델들을 마지막에 import (외래키 관계 있는 것들)
-from .users import UserEventSubscription, UserGoogleCalendar
-
-__all__ = [
-    "Base", 
-    "BaseModel", 
-    "Users", 
-    "LevelFeature", 
-    "Events", 
-    "EventWebhook", 
-    "UserEventSubscription", 
-    "UserGoogleCalendar"
-]
+__all__ = ["Events", "EventWebhook", "Base"]
